requirements.txt
python-telegram-bot==21.6
feedparser==6.0.11
APScheduler==3.10.4
requests==2.32.3
python-dotenv==1.0.1
PyYAML==6.0.2
sqlalchemy==2.0.32

aiohttp==3.9.5
Pillow==10.4.0
